feat: moving uds handling to decorator and registry
