+++ packages/pandemic-common/src/pandemic_common/__init__.py
@@ -2,6 +2,15 @@
 
 __version__ = "0.0.1"
 
+from .daemon import UnixDaemonServer
 from .events import EventClient, EventManager, EventSubscription
+from .routing import RouteRegistry, route
 
-__all__ = ["EventClient", "EventManager", "EventSubscription"]
+__all__ = [
+    "UnixDaemonServer",
+    "route",
+    "RouteRegistry",
+    "EventClient",
+    "EventManager",
+    "EventSubscription",
+]
