+++ packages/pandemic-core/src/pandemic_core/config.py
@@ -15,6 +15,7 @@ class DaemonConfig:
     socket_path: str = "/var/run/pandemic/daemon.sock"
     # Int conversation does a base 10 on a base 8 and back to base 8
     socket_mode: int = 660
+    socket_owner: str = "pandemic"
     socket_group: str = "pandemic"
     pid_file: str = "/var/run/pandemic/daemon.pid"
     infections_dir: str = "/opt/pandemic/infections"
@@ -61,6 +62,7 @@ class DaemonConfig:
         return cls(
             socket_path=daemon_config.get("socket_path", cls.socket_path),
             socket_mode=int(str(daemon_config.get("socket_mode", cls.socket_mode))),
+            socket_owner=daemon_config.get("socket_owner", cls.socket_owner),
             socket_group=daemon_config.get("socket_group", cls.socket_group),
             pid_file=daemon_config.get("pid_file", cls.pid_file),
             infections_dir=storage_config.get("infections_dir", cls.infections_dir),
@@ -82,6 +84,7 @@ class DaemonConfig:
         return cls(
             socket_path=os.getenv("PANDEMIC_SOCKET_PATH", cls.socket_path),
             socket_mode=int(os.getenv("PANDEMIC_SOCKET_MODE", str(cls.socket_mode))),
+            socket_owner=os.getenv("PANDEMIC_SOCKET_OWNER", cls.socket_owner),
             socket_group=os.getenv("PANDEMIC_SOCKET_GROUP", cls.socket_group),
             pid_file=os.getenv("PANDEMIC_PID_FILE", cls.pid_file),
             infections_dir=os.getenv("PANDEMIC_INFECTIONS_DIR", cls.infections_dir),
@@ -136,6 +139,7 @@ class DaemonConfig:
             "daemon": {
                 "socket_path": self.socket_path,
                 "socket_mode": oct(self.socket_mode),
+                "socket_owner": self.socket_owner,
                 "socket_group": self.socket_group,
                 "pid_file": self.pid_file,
             },
